web/requirements.txt
dash
dash-bootstrap-components
dash-extensions
loguru
flask
flask-sqlalchemy
flask-login
requests
dash-loading-spinners
python-dotenv
